Create module dragDrop.js
